refactor(utils): fix layering violation in branch_compare
Replace TYPE_CHECKING import of GitClient with a local Protocol
(_GitRunner) and import GitError from its canonical location
(vibe3.exceptions) instead of vibe3.clients.git_client.

This resolves the remaining utils → clients layering violation.
The other 7 violations were already fixed by PR #1361.

Changes:
- src/vibe3/utils/branch_compare.py: Use Protocol for dependency inversion
- tests/vibe3/utils/test_branch_compare.py: Update GitError import

All callers pass GitClient instances that satisfy the _GitRunner protocol.

Fixes #1578
